:truck: move `LOW_TIER_INFO` and `HIGH_TIER_INFO` constants
