+++ src/components/frame/with-buttons/menu-button/game-info/constants/index.ts
@@ -1,3 +1,4 @@
 export const INFO_TYPE = ['main', 'item', 'game', 'tier'] as const;
 
 export { ITEMS } from './item';
+export { LOW_TIER_INFO, HIGH_TIER_INFO } from './tier';
